refactor: Adjust variable and binds for chart and table
